# openapi/tickers.yaml
# Generated by `python -m scripts.export_openapi`. Do not edit by hand.
openapi: 3.1.0
info:
  title: 'Stock Movement Explainer: tickers'
  description: Explains major daily stock price movements using company, industry and macro news.
  version: 1.0.0
tags:
- name: tickers
  description: Ingest a ticker and read its prices, major movements, explanations and news.
paths:
  /tickers:
    get:
      tags:
      - tickers
      summary: List ingested tickers
      operationId: list_tickers_tickers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TickerSummaryResponse'
                type: array
                title: Response List Tickers Tickers Get
  /tickers/{ticker}/ingest:
    post:
      tags:
      - tickers
      summary: Fetch prices, detect major movements, find news and explain them (runs in the background)
      description: 'Returns the job to poll at `GET /tickers/{ticker}/status`.


        Safe to repeat: finished work is never redone or re-billed, and an ingest already in flight for the ticker is

        returned (200) rather than started twice. Missing API keys fail here with 503, before any job is created.'
      operationId: ingest_ticker_tickers__ticker__ingest_post
      parameters:
      - name: ticker
        in: path
        required: true
        schema:
          type: string
          title: Ticker
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/IngestRequest'
              - type: 'null'
              title: Request
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestJobResponse'
        '200':
          description: An ingest is already running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestJobResponse'
        '422':
          description: The ticker is not a valid symbol, or a parameter failed validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: An upstream provider (Exa or OpenAI) failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: A required API key is not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tickers/{ticker}/status:
    get:
      tags:
      - tickers
      summary: Latest ingest job for a ticker
      operationId: ingest_status_tickers__ticker__status_get
      parameters:
      - name: ticker
        in: path
        required: true
        schema:
          type: string
          title: Ticker
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestJobResponse'
        '404':
          description: Nothing has been ingested for this ticker
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: The ticker is not a valid symbol, or a parameter failed validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tickers/{ticker}:
    get:
      tags:
      - tickers
      summary: 'All stock and news data for a ticker: prices, major movements, explanations and articles'
      description: 'Movement-level filters (dates, direction, size, category, driver_hint, confidence) choose which movements
        are

        returned; article-level filters (`tier`, `min_relevance`) narrow the articles shown inside each movement.'
      operationId: get_ticker_tickers__ticker__get
      parameters:
      - name: ticker
        in: path
        required: true
        schema:
          type: string
          title: Ticker
      - name: start
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          description: Earliest movement date (inclusive)
          title: Start
        description: Earliest movement date (inclusive)
      - name: end
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          description: Latest movement date (inclusive)
          title: End
        description: Latest movement date (inclusive)
      - name: direction
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Direction'
          - type: 'null'
          description: Only gains (up) or only losses (down)
          title: Direction
        description: Only gains (up) or only losses (down)
      - name: min_abs_change
        in: query
        required: false
        schema:
          anyOf:
          - type: number
            minimum: 0
          - type: 'null'
          description: Minimum absolute daily move, in percent
          title: Min Abs Change
        description: Minimum absolute daily move, in percent
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/ExplanationCategory'
          - type: 'null'
          description: Explanation verdict
          title: Category
        description: Explanation verdict
      - name: driver_hint
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/DriverHint'
          - type: 'null'
          description: What the price data alone suggested
          title: Driver Hint
        description: What the price data alone suggested
      - name: min_confidence
        in: query
        required: false
        schema:
          anyOf:
          - type: number
            maximum: 1
            minimum: 0
          - type: 'null'
          description: Minimum explanation confidence
          title: Min Confidence
        description: Minimum explanation confidence
      - name: explained_only
        in: query
        required: false
        schema:
          type: boolean
          description: Drop movements that have no explanation yet
          default: false
          title: Explained Only
        description: Drop movements that have no explanation yet
      - name: tier
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/NewsTier'
          - type: 'null'
          description: Only show articles found by this search tier
          title: Tier
        description: Only show articles found by this search tier
      - name: min_relevance
        in: query
        required: false
        schema:
          anyOf:
          - type: number
            maximum: 1
            minimum: 0
          - type: 'null'
          description: Only show articles at or above this relevance
          title: Min Relevance
        description: Only show articles at or above this relevance
      - name: sort
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/MovementSort'
          description: Result order
          default: date_desc
        description: Result order
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: include_prices
        in: query
        required: false
        schema:
          type: boolean
          description: Include daily price bars (within start/end)
          default: true
          title: Include Prices
        description: Include daily price bars (within start/end)
      - name: include_news
        in: query
        required: false
        schema:
          type: boolean
          description: Include each movement's articles
          default: true
          title: Include News
        description: Include each movement's articles
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TickerDataResponse'
        '404':
          description: Nothing has been ingested for this ticker
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: The ticker is not a valid symbol, or a parameter failed validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tickers/{ticker}/movements/{day}:
    get:
      tags:
      - tickers
      summary: 'One movement in full: stats, explanation and every article considered'
      operationId: get_movement_tickers__ticker__movements__day__get
      parameters:
      - name: ticker
        in: path
        required: true
        schema:
          type: string
          title: Ticker
      - name: day
        in: path
        required: true
        schema:
          type: string
          format: date
          title: Day
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovementResponse'
        '404':
          description: The ticker is not ingested, or it has no major movement on that date
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: The ticker is not a valid symbol, or a parameter failed validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ArticleResponse:
      properties:
        id:
          type: integer
          title: Id
        url:
          type: string
          title: Url
        title:
          type: string
          title: Title
        source:
          anyOf:
          - type: string
          - type: 'null'
          title: Source
        published_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Published At
        snippet:
          anyOf:
          - type: string
          - type: 'null'
          title: Snippet
        tier:
          $ref: '#/components/schemas/NewsTier'
        relevance:
          anyOf:
          - type: number
          - type: 'null'
          title: Relevance
        cited:
          type: boolean
          title: Cited
      type: object
      required:
      - id
      - url
      - title
      - source
      - published_at
      - snippet
      - tier
      - relevance
      - cited
      title: ArticleResponse
    CompanyResponse:
      properties:
        ticker:
          type: string
          title: Ticker
        name:
          type: string
          title: Name
        sector:
          anyOf:
          - type: string
          - type: 'null'
          title: Sector
        industry:
          anyOf:
          - type: string
          - type: 'null'
          title: Industry
        sector_etf:
          anyOf:
          - type: string
          - type: 'null'
          title: Sector Etf
        peers:
          items:
            $ref: '#/components/schemas/PeerResponse'
          type: array
          title: Peers
      type: object
      required:
      - ticker
      - name
      - sector
      - industry
      - sector_etf
      - peers
      title: CompanyResponse
    Direction:
      type: string
      enum:
      - up
      - down
      title: Direction
      description: Sign of a move, for filtering.
    DriverHint:
      type: string
      enum:
      - market
      - sector
      - idiosyncratic
      title: DriverHint
      description: What the price data alone suggests drove a move (D4).
    ErrorBody:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
      type: object
      required:
      - code
      - message
      title: ErrorBody
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      type: object
      required:
      - error
      title: ErrorResponse
      description: 'Body of every error raised by the application (see `AppError`). Request validation errors use FastAPI''s

        standard 422 body instead.'
    ExplanationCategory:
      type: string
      enum:
      - company
      - industry
      - macro
      - unexplained
      title: ExplanationCategory
      description: The explanation pass's verdict on a move (D6). UNEXPLAINED is a valid answer.
    ExplanationResponse:
      properties:
        summary:
          type: string
          title: Summary
        category:
          $ref: '#/components/schemas/ExplanationCategory'
        confidence:
          type: number
          title: Confidence
        model:
          type: string
          title: Model
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - summary
      - category
      - confidence
      - model
      - created_at
      title: ExplanationResponse
    IngestJobResponse:
      properties:
        id:
          type: integer
          title: Id
        ticker:
          type: string
          title: Ticker
        status:
          $ref: '#/components/schemas/JobStatus'
        stage:
          anyOf:
          - type: string
          - type: 'null'
          title: Stage
        detail:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Detail
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
        started_at:
          type: string
          format: date-time
          title: Started At
        finished_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Finished At
      type: object
      required:
      - id
      - ticker
      - status
      - stage
      - detail
      - error
      - started_at
      - finished_at
      title: IngestJobResponse
    IngestRequest:
      properties:
        start:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Start
          description: 'First day to analyse. Default: `lookback_days` before `end`'
        end:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: End
          description: 'Last day to analyse. Default: today'
        lookback_days:
          anyOf:
          - type: integer
            maximum: 1825.0
            minimum: 1.0
          - type: 'null'
          title: Lookback Days
          description: Used when `start` is omitted
        threshold_pct:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Threshold Pct
          description: 'Major-movement cutoff in percent. Default: 2.0'
        max_movements:
          anyOf:
          - type: integer
            maximum: 200.0
            minimum: 1.0
          - type: 'null'
          title: Max Movements
          description: 'Cost guard: how many of the largest moves get news + explanation'
        refresh:
          type: boolean
          title: Refresh
          description: Re-explain movements that already have an explanation
          default: false
      type: object
      title: IngestRequest
      description: 'All fields optional: an empty body ingests the last year at the configured threshold.'
    JobStatus:
      type: string
      enum:
      - pending
      - running
      - done
      - failed
      title: JobStatus
      description: Lifecycle of an ingest job.
    MovementFilters:
      properties:
        start:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Start
          description: Earliest movement date (inclusive)
        end:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: End
          description: Latest movement date (inclusive)
        direction:
          anyOf:
          - $ref: '#/components/schemas/Direction'
          - type: 'null'
          description: Only gains (up) or only losses (down)
        min_abs_change:
          anyOf:
          - type: number
            minimum: 0.0
          - type: 'null'
          title: Min Abs Change
          description: Minimum absolute daily move, in percent
        category:
          anyOf:
          - $ref: '#/components/schemas/ExplanationCategory'
          - type: 'null'
          description: Explanation verdict
        driver_hint:
          anyOf:
          - $ref: '#/components/schemas/DriverHint'
          - type: 'null'
          description: What the price data alone suggested
        min_confidence:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: 0.0
          - type: 'null'
          title: Min Confidence
          description: Minimum explanation confidence
        explained_only:
          type: boolean
          title: Explained Only
          description: Drop movements that have no explanation yet
          default: false
        tier:
          anyOf:
          - $ref: '#/components/schemas/NewsTier'
          - type: 'null'
          description: Only show articles found by this search tier
        min_relevance:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: 0.0
          - type: 'null'
          title: Min Relevance
          description: Only show articles at or above this relevance
        sort:
          $ref: '#/components/schemas/MovementSort'
          description: Result order
          default: date_desc
        limit:
          type: integer
          maximum: 200.0
          minimum: 1.0
          title: Limit
          default: 50
        offset:
          type: integer
          minimum: 0.0
          title: Offset
          default: 0
      type: object
      title: MovementFilters
      description: 'The one definition of how movements can be filtered.


        Used as the REST query parameters, as the chat tools'' arguments (their JSON schema is generated from this model)

        and as the repository''s input, so the three stay in sync.'
    MovementResponse:
      properties:
        ticker:
          type: string
          title: Ticker
        date:
          type: string
          format: date
          title: Date
        pct_change:
          type: number
          title: Pct Change
        close:
          type: number
          title: Close
        prev_close:
          type: number
          title: Prev Close
        zscore:
          anyOf:
          - type: number
          - type: 'null'
          title: Zscore
        volume_ratio:
          anyOf:
          - type: number
          - type: 'null'
          title: Volume Ratio
        market_pct_change:
          anyOf:
          - type: number
          - type: 'null'
          title: Market Pct Change
        sector_pct_change:
          anyOf:
          - type: number
          - type: 'null'
          title: Sector Pct Change
        excess_vs_market:
          anyOf:
          - type: number
          - type: 'null'
          title: Excess Vs Market
        excess_vs_sector:
          anyOf:
          - type: number
          - type: 'null'
          title: Excess Vs Sector
        driver_hint:
          $ref: '#/components/schemas/DriverHint'
        news_window_start:
          type: string
          format: date
          title: News Window Start
        news_window_end:
          type: string
          format: date
          title: News Window End
        peer_moves:
          items:
            $ref: '#/components/schemas/PeerMoveResponse'
          type: array
          title: Peer Moves
        explanation:
          anyOf:
          - $ref: '#/components/schemas/ExplanationResponse'
          - type: 'null'
        articles:
          items:
            $ref: '#/components/schemas/ArticleResponse'
          type: array
          title: Articles
      type: object
      required:
      - ticker
      - date
      - pct_change
      - close
      - prev_close
      - zscore
      - volume_ratio
      - market_pct_change
      - sector_pct_change
      - excess_vs_market
      - excess_vs_sector
      - driver_hint
      - news_window_start
      - news_window_end
      - peer_moves
      - explanation
      - articles
      title: MovementResponse
    MovementSort:
      type: string
      enum:
      - date_desc
      - date_asc
      - magnitude
      title: MovementSort
    NewsTier:
      type: string
      enum:
      - company
      - industry
      - macro
      title: NewsTier
      description: Which search tier surfaced an article.
    PeerMoveResponse:
      properties:
        name:
          type: string
          title: Name
        ticker:
          type: string
          title: Ticker
        pct_change:
          type: number
          title: Pct Change
      type: object
      required:
      - name
      - ticker
      - pct_change
      title: PeerMoveResponse
    PeerResponse:
      properties:
        name:
          type: string
          title: Name
        ticker:
          anyOf:
          - type: string
          - type: 'null'
          title: Ticker
      type: object
      required:
      - name
      - ticker
      title: PeerResponse
    PriceResponse:
      properties:
        date:
          type: string
          format: date
          title: Date
        open:
          type: number
          title: Open
        high:
          type: number
          title: High
        low:
          type: number
          title: Low
        close:
          type: number
          title: Close
        volume:
          type: integer
          title: Volume
      type: object
      required:
      - date
      - open
      - high
      - low
      - close
      - volume
      title: PriceResponse
    TickerDataResponse:
      properties:
        summary:
          $ref: '#/components/schemas/TickerSummaryResponse'
        latest_ingest:
          anyOf:
          - $ref: '#/components/schemas/IngestJobResponse'
          - type: 'null'
        filters:
          $ref: '#/components/schemas/MovementFilters'
        total_movements:
          type: integer
          title: Total Movements
        movements:
          items:
            $ref: '#/components/schemas/MovementResponse'
          type: array
          title: Movements
        prices:
          anyOf:
          - items:
              $ref: '#/components/schemas/PriceResponse'
            type: array
          - type: 'null'
          title: Prices
      type: object
      required:
      - summary
      - latest_ingest
      - filters
      - total_movements
      - movements
      - prices
      title: TickerDataResponse
    TickerSummaryResponse:
      properties:
        company:
          $ref: '#/components/schemas/CompanyResponse'
        first_price_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: First Price Date
        last_price_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Last Price Date
        movement_count:
          type: integer
          title: Movement Count
        explained_count:
          type: integer
          title: Explained Count
      type: object
      required:
      - company
      - first_price_date
      - last_price_date
      - movement_count
      - explained_count
      title: TickerSummaryResponse